Validation refactor

+++ RestfulApi/Controllers/LanguagesController.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using RestfulApi.BusinessLayer.Abstract;
 using RestfulApi.DtoLayer.DTOs.LanguageDtos;
+using RestfulApi.EntityLayer.Entities;
+using RestfulApi.Validators.LanguageValidators;
 
 namespace RestfulApi.Controllers
 {
@@ -26,9 +28,17 @@ namespace RestfulApi.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetLanguageById(int id)
         {
+            var validator = new GetLanguageByIdValidator();
+            var validationResult = validator.Validate(id);
+
+            if(!validationResult.IsValid)
+                return BadRequest(validationResult.Errors.Select(error => error.ErrorMessage));
+
             var language = await _languageService.GetLanguageByIdAsync(id);
+
             if (language == null)
-                return NotFound();
+                return NotFound("Language not found !");
+
             return Ok(language);
         }
 
@@ -42,6 +52,13 @@ namespace RestfulApi.Controllers
         [HttpDelete]
         public async Task<IActionResult> DeleteLanguage(int id)
         {
+            var language = new Language { Id = id };
+            var validator = new DeleteLanguageValidator();
+            var validationResult = validator.Validate(language);
+
+            if(!validationResult.IsValid)
+                return BadRequest(validationResult.Errors.Select(error => error.ErrorMessage));
+
             await _languageService.DeleteLanguageAsync(id);
             return Ok("Language deleted successfully !");
         }
@@ -49,6 +66,12 @@ namespace RestfulApi.Controllers
         [HttpPut]
         public async Task<IActionResult> UpdateLanguage(UpdateLanguageDto updateLanguageDto)
         {
+            var validator = new UpdateLanguageValidator();
+            var validationResult = await validator.ValidateAsync(updateLanguageDto);
+
+            if (!validationResult.IsValid)
+                return BadRequest(validationResult.Errors.Select(error => error.ErrorMessage));
+
             await _languageService.UpdateLanguageAsync(updateLanguageDto);
             return Ok("Language updated successfully !");
         }
